langchain/ChatGroq/requirements.txt
streamlit
python-dotenv
langchain
langchain-groq
langchain-core
